final fixes to data loading.  Added wash covariate, added depth category, trimmed out the dilutions that should be excluded. Ensured all of the observations had lat-lons.
Also added the ability to specify covariates within the "format_qPCR_data" function (continuous, factors, and smoothes).  The smoothes does not report the proper objects as yet, so do not use it.  But it makes the correct matrices.This commit breaks the indexing for plateSample and splits the standards and the unknown samples.